Move functional harness to a separate module to avoid import errors.
